Rename fluid_sf2 to fluid_sffile
